MINOR: Move common consumer tests out of abstract consumer class
